+++ src/flows/Termination/json-schemas/terminationDetails.ts
@@ -117,6 +117,14 @@ export const terminationDetailsSchema = {
             multiple: true,
           },
         },
+        risk_assesment_info: {
+          description: '',
+          title: 'Risk assessment info',
+          type: 'null',
+          'x-jsf-presentation': {
+            inputType: 'hidden',
+          },
+        },
         risk_assessment_reasons: {
           description: '',
           title: 'This employee is...',
@@ -225,6 +233,7 @@ export const terminationDetailsSchema = {
         'reason_description',
         'additional_comments',
         'termination_reason_files',
+        'risk_assesment_info',
         'risk_assessment_reasons',
         'will_challenge_termination',
         'will_challenge_termination_description',
